// examples/chat-app/source-starter-for-tutorial/frontend/src/lib/api/models/counterparty-dto.ts
/*
 * Copyright 2025 sovity GmbH
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 *
 * Contributors:
 *     sovity - init and continued development
 */
import {z} from 'zod';
import {ConnectionStatusDto} from '@/lib/api/models/connection-status-dto';

const CounterpartySchema = z.object({
  participantId: z.string(),
  connectorEndpoint: z.string().url(),
  status: z.nativeEnum(ConnectionStatusDto),
  lastUpdate: z.date(),
});

export type CounterpartyDto = z.infer<typeof CounterpartySchema>;